Moved claim.go into the token package
